+++ scripts/defs.sh
@@ -28,40 +28,55 @@ ctrl_c_handler() {
 # --- Application-specific functions ---
 # --------------------------------------
 
-# Generate latest API files, copying from one directory to another.
-#
-# Note: do not pass directories that resolve to the same path (i.e. they point
-# to the same location), otherwise data will be lost (rm command used).
-gen_api_files() {
+# Copies latest API files from one directory to another. The destination
+# directory will be created if it doesn't exists.
+copy_api_files() {
     if [ "$#" -ne 2 ]; then
-        echo "2 parameters expected to generate dependencies, $# given" 1>&2 && return 1
+        echo "2 parameters expected to copy API files, $# given" 1>&2 && return 1
     fi
 
-    local src_dir jsu_dir nvc_dir dst_dir
-    src_dir="$1"
-    jsu_dir="$src_dir/node_modules/jsupack/src"
-    nvc_dir="$src_dir/src"
+    local base_dir jsu_dir nvc_dir dst_dir
+    base_dir="$1"
+    jsu_dir="$base_dir/node_modules/jsupack/src"
+    nvc_dir="$base_dir/src"
     dst_dir="$2"
 
-    echo "Generating latest API files..."
     if [[ ! -d "$jsu_dir" ]]; then
-        echo "$jsu_dir is not a directory" 1>&2 && return 1
+        echo "jsu path '$jsu_dir' is not a directory" 1>&2 && return 1
     fi
     if [[ ! -d "$nvc_dir" ]]; then
-        echo "$nvc_dir is not a directory" 1>&2 && return 1
+        echo "nvc path '$nvc_dir' is not a directory" 1>&2 && return 1
     fi
 
-    echo "Delete and recreate the destination directory ($dst_dir)" &&
-    rm -rf "$dst_dir" && mkdir -p "$dst_dir" || return 1
-
-    echo "Copy dependencies to destination directory" &&
-    cp "$jsu_dir/jsu_common.js" "$dst_dir/jsu_common.js" &&
+    mkdir -p "$dst_dir" &&
+    cp "$jsu_dir/jsu_common.js"     "$dst_dir/jsu_common.js" &&
     cp "$jsu_dir/jsu_csv_parser.js" "$dst_dir/jsu_csv_parser.js" &&
-    cp "$jsu_dir/jsu_event.js" "$dst_dir/jsu_event.js" &&
-    cp "$jsu_dir/jsu_latex.js" "$dst_dir/jsu_latex.js" &&
-    cp "$nvc_dir/nvc.js" "$dst_dir/nvc.js" &&
-    cp "$nvc_dir/nvc_fsm.js" "$dst_dir/nvc_fsm.js" &&
-    cp "$nvc_dir/nvc_quick.js" "$dst_dir/nvc_quick.js" || return 1
+    cp "$jsu_dir/jsu_event.js"      "$dst_dir/jsu_event.js" &&
+    cp "$jsu_dir/jsu_latex.js"      "$dst_dir/jsu_latex.js" &&
+    cp "$nvc_dir/nvc.js"            "$dst_dir/nvc.js" &&
+    cp "$nvc_dir/nvc_fsm.js"        "$dst_dir/nvc_fsm.js" &&
+    cp "$nvc_dir/nvc_quick.js"      "$dst_dir/nvc_quick.js" || return 1
+
+    return 0
+}
+
+# Updates HTML examples with latest dependencies.
+update_html_examples() {
+    if [ "$#" -ne 1 ]; then
+        echo "1 parameter expected to update HTML examples, $# given" 1>&2 && return 1
+    fi
+
+    local base_dir examples_dir api_files_dir
+    base_dir="$1"
+    examples_dir="$base_dir/examples/html_pages"
+    api_files_dir="$examples_dir/api_files"
+
+    echo "Updating HTML examples..."
+
+    echo "Delete API files directory '$api_files_dir'" &&
+    rm -rf "$api_files_dir" &&
+    echo "Copy latest API files to their directory" &&
+    copy_api_files "$base_dir" "$api_files_dir" || return 1
 
     echo "Finished!"
     return 0